also change interface here
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4065 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/wpc/dmddevice.cpp
@@ -31,10 +31,10 @@ Set_4_Colors_Palette_t DmdDev_Set_4_Colors_Palette;
 typedef void (*Console_Data_t)(UINT8 data);
 Console_Data_t DmdDev_Console_Data;
 
-typedef void (*Render_16_Shades_t)(UINT8 width, UINT8 height, UINT8 *currbuffer);
+typedef void (*Render_16_Shades_t)(UINT16 width, UINT16 height, UINT8 *currbuffer);
 Render_16_Shades_t DmdDev_Render_16_Shades;
 
-typedef void (*Render_4_Shades_t)(UINT8 width, UINT8 height, UINT8 *currbuffer);
+typedef void (*Render_4_Shades_t)(UINT16 width, UINT16 height, UINT8 *currbuffer);
 Render_4_Shades_t DmdDev_Render_4_Shades;
 
 typedef void (*render_PM_Alphanumeric_Frame_t)(layout_t layout, const UINT16 *const seg_data, const UINT16 *const seg_data2);
